feat: refactor kotlin
