+++ packages/kotlin/flappy/src/main/kotlin/flappy/ChatMessage.kt
@@ -1,24 +1,24 @@
 package flappy
 
 enum class FlappyRole {
-    SYSTEM, USER, ASSISTANT
+  SYSTEM, USER, ASSISTANT
 }
 
 abstract class FlappyChatMessage(open val content: String) {
-    abstract val role: FlappyRole
+  abstract val role: FlappyRole
 
-    override fun toString() = "[$role] $content"
+  override fun toString() = "[$role] $content"
 }
 
 class SystemMessage(override val content: String) : FlappyChatMessage(content) {
-    override val role = FlappyRole.SYSTEM
+  override val role = FlappyRole.SYSTEM
 
 }
 
 class UserMessage(override val content: String) : FlappyChatMessage(content) {
-    override val role = FlappyRole.USER
+  override val role = FlappyRole.USER
 }
 
 class AssistantMessage(override val content: String) : FlappyChatMessage(content) {
-    override val role = FlappyRole.ASSISTANT
+  override val role = FlappyRole.ASSISTANT
 }
