+++ packages/kotlin/flappy/src/main/kotlin/flappy/Constant.kt
