+++ packages/kotlin/flappy/src/main/kotlin/flappy/Exception.kt
@@ -11,3 +11,5 @@ class FatalException(message: String) : Exception(message)
 class RuntimeException(message: String) : Exception(message)
 
 class ParseException(message: String) : Exception(message)
+
+class FieldNotFoundException(message: String) : Exception(message)
