+++ packages/kotlin/flappy/src/main/kotlin/flappy/LLM.kt
@@ -1,0 +1,39 @@
+package flappy
+
+import java.util.logging.Logger
+
+interface FlappyLLM {
+  interface GenerateConfig {
+    val maxTokens: Int
+    val temperature: Double
+    val topP: Double
+  }
+
+  abstract class Response(open val data: String) {
+    abstract val success: Boolean
+
+    override fun toString() = "[$success] $data"
+  }
+
+  class SuccessLLMResponse(override val data: String) : Response(data) {
+    override val success = true
+
+  }
+
+  class FailLLMResponse(override val data: String) : Response(data) {
+    override val success = false
+  }
+}
+
+
+abstract class FlappyLLMBase {
+  open val defaultMaxTokens: Int = 0
+  protected val logger: Logger = Logger.getLogger(this.javaClass.name)
+
+  abstract suspend fun chatComplete(
+    messages: List<FlappyChatMessage>,
+    source: String,
+    config: FlappyLLM.GenerateConfig? = null
+  ): FlappyLLM.Response
+
+}
