--- packages/kotlin/flappy/src/main/kotlin/ktflappy/Reflect.kt
@@ -1,13 +1,0 @@
-package flappy
-
-import kotlin.reflect.KProperty1
-
-//https://stackoverflow.com/a/35539628/20030734
-@Suppress("UNCHECKED_CAST")
-fun <R> readInstanceProperty(instance: Any, propertyName: String): R {
-    val property = instance::class.members
-        // don't cast here to <Any, R>, it would succeed silently
-        .first { it.name == propertyName } as KProperty1<Any, *>
-    // force a invalid cast exception if incorrect type here
-    return property.get(instance) as R
-}
