--- packages/kotlin/flappy/src/main/kotlin/ktflappy/llms/Dummy.kt
@@ -1,14 +1,0 @@
-package flappy.llms
-
-import flappy.FlappyChatMessage
-import flappy.FlappyLLM
-import flappy.FlappyLLMBase
-
-class Dummy : FlappyLLMBase() {
-    override suspend fun chatComplete(
-        messages: List<FlappyChatMessage>,
-        config: FlappyLLM.GenerateConfig?
-    ): FlappyLLM.Response {
-        return FlappyLLM.SuccessLLMResponse("")
-    }
-}
