+++ packages/kotlin/flappy/src/test/java/flappy/AgentTestJava.java
@@ -1,6 +1,5 @@
 package flappy;
 
 public class AgentTestJava {
-
-
+  
 }
