+++ packages/kotlin/flappy/src/test/java/flappy/examples/ResumeTestJava.java
@@ -1,0 +1,4 @@
+package flappy.examples;
+
+public class ResumeTestJava {
+}
