+++ packages/kotlin/flappy/src/test/kotlin/flappy/LibraryTest.kt
@@ -1,0 +1,62 @@
+/*
+ * This Kotlin source file was generated by the Gradle 'init' task.
+ */
+package flappy
+
+import com.fasterxml.jackson.annotation.JsonInclude
+import com.fasterxml.jackson.module.kotlin.readValue
+import kotlin.test.Test
+
+private class LibraryTest {
+
+  @Test
+  fun github437() {
+    data class KmsCmkSecret(val cmkIdentifier: String, val dataKeyCiphertext: String)
+
+    val secretJson = """{"cmkIdentifier": "foo", "dataKeyCiphertext":"bar"}"""
+    val kmsCmkSecret1 = jacksonMapper.readValue<KmsCmkSecret>(secretJson)
+    println(kmsCmkSecret1)
+    val kmsCmkSecret2: KmsCmkSecret = jacksonMapper.readValue(secretJson)
+    println(kmsCmkSecret2)
+
+    val secretArrayJSON = """[]"""
+    val kmsCmkSecretArray = jacksonMapper.readValue<List<KmsCmkSecret>>(secretArrayJSON)
+    println(kmsCmkSecretArray)
+  }
+  
+  @Test
+  fun xxx() {
+    data class Foo(val xxx: String)
+
+    data class Bar(val foo: Foo)
+
+    data class Baz(val bar: List<Bar>, val foo: Foo)
+
+    val foo = Foo(xxx = "foo")
+    val baz = Baz(bar = listOf(Bar(foo = foo)), foo = foo)
+
+    println(jacksonMapper.writeValueAsString(baz))
+    println(jacksonMapper.writeValueAsString(listOf(baz)))
+
+    @JsonInclude(JsonInclude.Include.NON_EMPTY)
+    data class Opt(
+      val xxx: String? = null,
+
+      val y: String = "123",
+
+      val obj: Any? = object {
+        val answer = "123"
+      },
+
+      val sss: String = "",
+
+      val none: Any? = null,
+
+      val map: Map<String, Int> = mapOf(Pair("foo", 1), Pair("bar", 2))
+    ) {
+      val foobar = "123132"
+    }
+    println(jacksonMapper.writeValueAsString(Opt()))
+  }
+
+}
