+++ packages/kotlin/flappy/src/test/kotlin/flappy/examples/ResumeTest.kt
@@ -1,0 +1,4 @@
+package flappy.examples
+
+class ResumeTest {
+}
